switch to kristaps@' oconfigure
It's a bit ugly to depend _only_ on the OpenBSD make infrastructure;
it's better when things are portable!  Making mymenu-portable would be
more effort than what it's worth it, so just make the main version
portable (again).  kristaps@' oconfigure (used also by mandoc-portable
and mymenu) is a lightweight configure system that's really nice to use.

see GitHub issue #1